Fix comparison of different signs

+++ PSTCollectionView/PSTCollectionView.m
@@ -1664,8 +1664,8 @@ static void PSTCollectionViewCommonSetup(PSTCollectionView *_self) {
         if (key.type == PSTCollectionViewItemTypeCell) {
             NSInteger oldGlobalIndex = [_update[@"oldModel"] globalIndexForItemAtIndexPath:key.indexPath];
             NSArray *oldToNewIndexMap = _update[@"oldToNewIndexMap"];
-            NSInteger newGlobalIndex = NSNotFound;
-            if (oldGlobalIndex >= 0 && oldGlobalIndex < [oldToNewIndexMap count]) {
+            NSUInteger newGlobalIndex = NSNotFound;
+            if (oldGlobalIndex >= 0 && (NSUInteger)oldGlobalIndex < [oldToNewIndexMap count]) {
                 newGlobalIndex = [oldToNewIndexMap[oldGlobalIndex] intValue];
             }
             NSIndexPath *newIndexPath = newGlobalIndex == NSNotFound ? nil : [_update[@"newModel"] indexPathForItemAtGlobalIndex:newGlobalIndex];
@@ -1880,7 +1880,7 @@ static void PSTCollectionViewCommonSetup(PSTCollectionView *_self) {
         }
     }
 
-    for(NSInteger i=0; i<[sortedInsertMutableItems count]; i++) {
+    for(NSUInteger i=0; i<[sortedInsertMutableItems count]; i++) {
         PSTCollectionViewUpdateItem *insertItem = sortedInsertMutableItems[i];
         NSIndexPath *indexPath = insertItem.indexPathAfterUpdate;
 
@@ -1899,7 +1899,7 @@ static void PSTCollectionViewCommonSetup(PSTCollectionView *_self) {
                 }
             }
 
-            NSInteger j=i+1;
+            NSUInteger j=i+1;
             while(j<[sortedInsertMutableItems count]) {
                 PSTCollectionViewUpdateItem *nextInsertItem = sortedInsertMutableItems[j];
 
@@ -2060,9 +2060,9 @@ static void PSTCollectionViewCommonSetup(PSTCollectionView *_self) {
     for(NSInteger i=0; i < [_collectionViewData numberOfItems]; i++)
         [newToOldMap addObject:@(NSNotFound)];
 
-    for(NSInteger i=0; i < [newModel count]; i++) {
+    for(NSUInteger i=0; i < [newModel count]; i++) {
         NSMutableArray* section = newModel[i];
-        for(NSInteger j=0; j<[section count];j++) {
+        for(NSUInteger j=0; j<[section count];j++) {
             NSInteger newGlobalIndex = [_collectionViewData globalIndexForItemAtIndexPath:[NSIndexPath indexPathForItem:j inSection:i]];
             if([section[j] intValue] != NSNotFound)
                 oldToNewMap[[section[j] intValue]] = @(newGlobalIndex);
